関数のモジュール化 #25
