+++ .gitignore
@@ -7,5 +7,4 @@ test
 icon
 
 # token
-token.json
-test_token.json
+token
